CRAYSAT-1593: Fix type-checking issues
This change fixes the type signature of the set_timeout() method so that
it permits `None` to be passed. This allows `None` to be used for the
`timeout` parameter for `requests.request()`, which does not time out
requests after a time limit.

Also add a `py.typed` file to comply with PEP 561.

Test Description:
Run type checker on SAT codebase after updating `csm-api-client` builds.