CRAYSAT-1620: Use authentication when accessing csm-python-modules
This commit adds creation of a .netrc file on the Jenkins host on which the
build is running. The intention of this file is to provide credential
authentication in the case when the package's dependencies are being installed
in a VirtualEnv on the host to facilitate running unit tests. This uses the
utility jenkinsUtils.writeNetRc() found in csm-shared-library.

Test Description: Built package in Jenkins