CRAYSAT-1650: Fix patch method of APIGatewayClient
The `_make_req` method was dropping the `json` kwarg when making the
request, which caused calls to the `patch` method that used the `json`
keyword argument to fail.

Test Description:
Unit tests pass. Manually made this change in a `sat bash` container
image and verified that the image renaming step in `sat bootprep` was
working again.