CRAYSAT-1330: Add support for naming images during customization
Add support to the `CFSClient` for specifying the name of the IMS image
during image customization CFS configuration sessions.

Add a method for getting the API version and a boolean cached property
which tells the caller whether the currently available CFS version
supports the specification of the resulting image name when creating an
image customization CFS configuration session.

Test Description:
Not tested yet.