CRAYSAT-1678: Fix token fetch during authentication
This change fixes the way tokens are handled when a new token is being
fetched. Previously when a token file didn't already exist for a token,
a new token would be fetched but not saved, and subsequent API calls
could not use the token.

Test Description:
Run unit tests. Test `sat auth` with empty config directory to ensure
that new tokens can be created.