CRAYSAT-1750: Update to newer poetry version
Update the build pipeline to use a newer version of poetry. This should
be a non-disruptive change, and it's good to stay up to date. This is
the latest version which was installed on my macbook when I installed
following the official install instructions.

Also update the `pyproject.toml` file to use the new way of specifying
additional package sources. See documentation of repositories here:

https://python-poetry.org/docs/repositories/#supplemental-package-sources

Test Description:
Installed the package with `poetry install` locally on my macbook.