CRAYSAT-1750: Allow PyYAML >= 6.0
Update version constraints in pyproject.toml to allow PyYAML >= 6.0.
This allows packages which require PyYAML 6.0.1 to address the Cython
3.0 incompatibility to continue building.

The main changes between PyYAML 5.x and 6.x are:

* `yaml.load` always requires a `Loader` argument
* Python 2.7 is no longer supported
* `YAMLLoadWarning` is gone

Only this last one affects this code. This is fixed by dropping the code
that catches the `YAMLLoadWarning`. We no longer need to catch this
warning anymore. The `kubernetes` library has not emitted that warning
since v10.0.0, and it certainly won't if installed with PyYAML 6.0.1.

Test Description:
Unit tests pass. Will be tested in `sat` before inclusion in a release.