CASMINST-6597: Add option to back up when saving CFSConfiguration
Add the ability to back up existing files or CFS configurations when saving
`CFSConfiguration` objects with `save_to_cfs` or `save_to_file` methods.
This is supported through a new `backup_suffix` string argument which
specifies the suffix to use when backup up the CFS configuration or the
file containing the CFS configuration payload. If the `backup_suffix` is
not specified, then it defaults to not backing up, which preserves
existing behavior.

Test Description:
New unit tests are passing.

Tested manually on my macbook in a Python virtual environment with the
`sat` package and this version of `csm_api_client` installed. Created a
`CFSClient` with `CFSClient(SATSession())` and then got a
`CFSConfiguration` object and tried various calls to `save_to_cfs` and
`save_to_file` methods.

Tested ability to create backups with `cfs-config-util` with new
`--create-backups` option on mug.