CASMINST-6597: Add methods to CFSClient to support cfs-config-util
Add a couple methods to the `CFSClient` class to support functionality
needed by `cfs-config-util`.

Add a method `get_component_ids_using_config` that gets a list of the
CFS component IDs for all components that have the given CFS
configuration assigned as their `desiredConfig`.

Add a method `update_component` that updates a CFS component based on
the given parameters.

Test Description:
New unit tests are passing.

Tested manually on my macbook in a Python virtual environment with the
`sat` package and this version of `csm_api_client` installed. Created a
`CFSClient` with `CFSClient(SATSession())` and then got a
`CFSConfiguration` object and called `get_component_ids_using_config`
with a few different configuration values and called `update_component`
with just `clear_error=True` to make sure it worked.

Tested by building a `cfs-config-util` that uses this new version of the
library from this branch and executed it on mug. Tested that it
correctly identified components which had their CFS configuration
updated and could also update component fields with various options.