CRAYSAT-1848: Fix kubernetes version constraint
The kubernetes package version constraint was requiring too new of a
version of kuberenetes. When this library is used by a version of sat
for CSM 1.5, we want to use a 22.y.z version of the kubernetes package
to match the 1.22 version of the Kubernetes cluster. When used by a
version of sat for CSM 1.6, we want to use a 23.y.z version of the
kubernetes package to match the 1.23 version of the Kubernetes cluster.

Relax the constraint to just require version 22 or newer. The lock file
can stay the same as it's not used when this package is included as a
dependency.