CRAYSAT-1330: Allow GET on `base_resource_path`
Modify the behavior of `APIGatewayClient._make_req` to construct the API
endpoint path such that it is possible to call the API at the
`base_resource_path` without a trailing `/` character.

This change in behavior is needed to allow us to query the CFS API at
the path `/cfs/v2` (and `/cfs/v3` in the near future). Queries with a
trailing slash get a 404 error.

Technically, this is a backwards-incompatible change because any users
of this library that were depending on a request without a path to make
a request on the `base_resource_path` _with_ the trailing slash will now
make the request with the trailing slash removed. It is still possible
to make the request on the `base_resource_path` _with_ the trailing
slash, but it must be done by passing an empty string as an argument, as
shown in the unit tests.

Test Description:
Unit tests only. Will be tested on a system in conjunction with a sat
change.