CRAYSAT-1330: Add semver requirement and regenerate poetry.lock
Added the `semver` package needed for the CFS version checking code, and
then ran `poetry lock --no-updates` to update the lock file. This took
around 90 minutes for some reason relating to the many many versions of
boto3 and botocore causing the poetry dependency resolver a lot of
trouble. I tried updating the poetry version to see if it would be
better, but it was not. Keeping the poetry version up to date seems good
regardless.