CRAYSAT-1840: Allow passing request parameters for PUT/PATCH/POST
We were mistakenly treating request parameters as the body of the
request for POST, PUT, and PATCH HTTP methods when they can also have
request parameters in addition to a request body.

The request parameters are needed now in order to pass a
`drop_branches=True` parameter while also passing in data in the request
body in a PUT request to CFS /configurations when creating a CFS
configuration.

This change should be backwards compatible as far as the public methods
of the `APIGatewayClient` class are concerned. It does not change the
behavior of any callers to the `get`, `put`, `patch`, or `post` methods.
It just ensures that their request body and parameters are passed to
`_make_req` appropriately.