CRAYSAT-1840: Add support for CFS v3 API
Add support for the CFS v3 API while preserving support for the CFS v2
API. This is accomplished by making CFS-API-version specific versions of
several of the classes in the `cfs` module. For backwards compatibility,
the existing unversioned class names, `CFSConfiguration`.
`CFSConfigurationLayer`, and `CFSClient` are now aliases for the
`CFSV2Configuration`, `CFSV2ConfigurationLayer`, and `CFSV2Client`
classes.

The `CFSImageConfigurationSession` class is still used for both CFS v2
and v3 since there were no large changes required, and it could be
easily generalized to both CFS API versions.

As much as possible, functionality that is common between CFS v2 and v3
is implemented in common abstract base classes which are inherited for
creating the version-specific classes.

This change also adds support for additional properties of CFS
configurations and configuration layers that were not recognized and
were discarded by old versions of this library. This includes:

* The `additional_inventory` property of a CFS configuration
* The `special_parameters` (or `specialParameters`) of a CFS
  configuration layer.
* Any additional data not understood by this library in a CFS
  configuration should now be retained when the configurations are
  loaded in via the library and then saved to CFS.

Support for handling paging of CFS components is now handled in the
`CFSV3Client` with a new `get_components` method.

This should be a fully backwards compatible upgrade of this library
since the interfaces to these classes have not changed.