CRAYSAT-1896: Add `CFSClient` methods to get configurations, sessions
Add methods to the `CFSV2Client` and `CFSV3Client` classes to get a list
of all the CFS configurations and sessions. The CFS V3 version handles
paging in the same way paging is handled for the `components` resource.

This change is needed for the `sat bootprep` command to be updated to
support CFS v2 and v3 because it looks through all the existing CFS
configurations to determine if there are any existing CFS configurations
of the same name.

Test Description:
Unit tests only so far. This will be tested by pulling it into the
corresponding branch in the sat repository that adds CFS v2/v3 support
to `sat bootprep`.