CRAYSAT-1896: Fix `CFSLayerBase.req_payload` to include False values
Fix the `req_payload` property of the `CFSLayerBase` class to include
values even when they are `False` (or Falsey). This ensures that the
`ims_require_dkms` parameter is included in the request payload, even
when set to `False`.

Add a unit test for this scenario.

Test Description:
Included in a build of `cray-sat` and tested `sat bootprep` on rocket.

(cherry picked from commit 8a3a5a476173fe2d393b0cbedfba619c56ea72e8)